Move evaluator into repl folder
